requirements.txt
anyio==4.3.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.2.2
Django==4.2.10
django-environ==0.11.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
python-decouple==3.8
python-telegram-bot==20.4
sniffio==1.3.0
sqlparse==0.4.4
typing-extensions==4.9.0
mysqlclient==2.1.1
Pillow==9.3.0
python-telegram==0.18.0
telegram-text==0.1.2
